Refactor test endpoints, implement of/on og image
